pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "flowback"
version = "0.1.0"
description = "Pick up exactly where you left off."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "google-generativeai>=0.7.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
]

[project.scripts]
flowback = "flowback.cli:app"

[tool.setuptools.packages.find]
include = ["flowback*"]
